Adds options to allow views to transition in parallel
EXPERIMENTAL:  By default, PinballY loads each view sequentially,
with playfield first, followed by Backglass, DMD, Topper, and finally
Instruction cards.  This approach results in smoother individual
transitions, as all the videos aren't trying to load at once.  The
side effect of this approach is that there is a perceptible delay
between views loading, particularly between when the playfield is
observed to load, and when the topper and instruction card are
are observed to load at the end of the sequennce.

This commit adds a "LoadViewsInParallel" config item that, when set,
will override this sequential behavior and begin loading all secondary
views in parallel as soon as the transition anmiation for the playfield
starts.  This config value defaults to false, retaining the original
behaviour if it is not set, or set to 0.

Note that this option has not been added to the in game options dialog.
To enable, you must directly add the following entry to settings.txt file:

LoadViewsInParallel = 1

This commit also adds the ability to tweak the transition times for the
playfield and secondary view fade animations. Adjusting these values can
help the timing of fades between playfield and secondary views to blend
a bit better when used in conjunction with the LoadViewsInParallel setting.

Again, these options have not been added to the in game options dialog.
To enable, you must add the following lines directly to settings.txt:

PlayfieldCrossfadeTime = 120
SecondaryCrossfadeTime = 120

These values represent the time in ms for the playfield and secondary
views to crossfade between videos or images.  If not specified, the
default setting is 120ms, retaining the exisiting transistion behavior.