Add ability to change weel layout parameters via Settings.txt
Adds the wheel radius, center location, angle of images, image size,
and location of the selected (n=0) game to Settings.txt. This allows
more tweaking of the UI.

For example, a nearly flat, bar-style wheel is now available with:

```
Wheel.Angle = 0.025
Wheel.Radius = 5.0
Wheel.YCenter = -5.25
```

NOTE: This commit adds but does not use the X coordinate settings.
NOTE: This has not been exported to the Javascript engine in any way.